Avoid a deadlock in the copy engine example
Because the reorder buffer moved from the PIM to the PCIe SS, a new deadlock
case appeared. When an AFU loops read responses back as writes, read responses
must have homes to avoid conflicts with writes. Use the new BUFFER_READ_RESPONSES
PIM parameter to guarantee that buffers are added when needed.